+++ CHANGES.md
@@ -10,6 +10,8 @@ Changes in v1.1b1
 Changes in v1.0.2
 -----------------
 
+- The Set-Printer-Attributes operation now properly validates the values of
+  "xxx-default" attributes (Issue #93)
 - The `papplSystemSetVersions` function now allows changes while the system is
   running (Issue #123)
 - The printing defaults page no longer shows a media chooser when there is a